22251226黄建龙/Project01/requirements.txt
numpy==1.23.5
pyopengl==3.6.1